backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.1
python-multipart>=0.0.9
langchain>=0.2.6
langchain-community>=0.2.6
supabase>=2.5.0
pypdf>=4.2.0
httpx>=0.27.0
langchain-google-genai>=1.0.6
python-dotenv>=1.0.1
Pillow>=10.3.0
